frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
python-docx==1.1.0